Possibly Fixed Polling Bug
