+++ packages/changed-elements-react/src/widgets/comparisonJobWidget/common/versionComapreToasts.ts
@@ -1,0 +1,67 @@
+import { IModelApp, NotifyMessageDetails, OutputMessagePriority, OutputMessageType } from "@itwin/core-frontend";
+import { NamedVersion } from "../../../clients/iModelsClient";
+import { toaster } from "@itwin/itwinui-react";
+import { ManagerStartComparisonV2Args, runManagerStartComparisonV2 } from "./versionCompareV2WidgetUtils";
+
+export const toastComparisonJobProcessing = (currentVersion: NamedVersion, targetVersion: NamedVersion) => {
+  IModelApp.notifications.outputMessage(
+    new NotifyMessageDetails(
+      OutputMessagePriority.Info,
+      IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.versionPickerTitle"),
+      `${IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.iModelVersions")}
+ <${currentVersion?.displayName}> ${IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.and")} <${targetVersion.displayName}>
+ ${IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.jobProcessing")}`,
+      OutputMessageType.Toast,
+    ),
+  );
+};
+
+export const toastComparisonJobError = (currentVersion: NamedVersion, targetVersion: NamedVersion) => {
+  IModelApp.notifications.outputMessage(
+    new NotifyMessageDetails(
+      OutputMessagePriority.Error,
+      IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.versionPickerTitle"),
+      `${IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.jobError")}
+ ${IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.iModelVersions")}
+ <${currentVersion?.displayName}> ${IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.and")} <${targetVersion.displayName}>`,
+      OutputMessageType.Toast,
+    ),
+  );
+};
+
+export const toastComparisonJobComplete = (args: ManagerStartComparisonV2Args) => {
+  const title = IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.viewTheReport");
+  toaster.closeAll();
+  toaster.setSettings({
+    placement: "bottom",
+  });
+  toaster.positive(
+    `${IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.iModelVersions")}<${args.currentVersion?.displayName}> ${IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.and")} <${args.targetVersion.displayName}> ${IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.jobComplete")}`, {
+    hasCloseButton: true,
+    link: {
+      title: title,
+      onClick: () => {
+        toaster.closeAll();
+        void runManagerStartComparisonV2({
+          comparisonJob: args.comparisonJob,
+          comparisonJobClient: args.comparisonJobClient,
+          iModelConnection: args.iModelConnection,
+          targetVersion: args.targetVersion,
+          currentVersion: args.currentVersion,
+        });
+      },
+    },
+    type: "persisting",
+  });
+};
+
+export const toastComparisonVisualizationStarting = () => {
+  IModelApp.notifications.outputMessage(
+    new NotifyMessageDetails(
+      OutputMessagePriority.Info,
+      IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.versionPickerTitle"),
+      IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.versionComparisonStarting"),
+      OutputMessageType.Toast,
+    ),
+  );
+};
